% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomProjection.R
\name{GenerateGaussianRandomProj}
\alias{GenerateGaussianRandomProj}
\title{gaussian_random_proj generates the dimension reduction "rotation" which is to be applied to
the numeric matrix of interest.}
\usage{
GenerateGaussianRandomProj(row, col)
}
\arguments{
\item{row}{is the number of rows in this rotation matrix.}

\item{col}{is the number of columns for this rotation matrix.}
}
\value{
the rotation matrix of dimensions \code{row} by \code{col}
}
\description{
gaussian_random_proj generates the dimension reduction "rotation" which is to be applied to
the numeric matrix of interest.
}

