% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomProjection.R
\name{JohnsonLindenstraussMinDim}
\alias{JohnsonLindenstraussMinDim}
\title{\code{johnson_loindenstrauss_min_dim} estimated the minimal size of the
random subspace to guarantee a bounded distortion introduced by the random projection.}
\usage{
JohnsonLindenstraussMinDim(n_samples, eps = 0.1)
}
\arguments{
\item{n_samples}{is the size of the subspace}

\item{eps}{the epsilon error which is acceptable.}
}
\value{
the conservative estiamtion of the minimal size of the random subspace give eps.
}
\description{
\code{johnson_loindenstrauss_min_dim} estimated the minimal size of the
random subspace to guarantee a bounded distortion introduced by the random projection.
}

