% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomProjection.R
\name{RandomProjection}
\alias{RandomProjection}
\title{gaussian_random_projection applies the projection onto the data.frame or matrix with the given
dimension or epsilon parameter.}
\usage{
RandomProjection(A, n_features = NULL, eps = 0.1)
}
\arguments{
\item{A}{is the data.frame or matrix which is of interest}

\item{n_features}{is the number of dimensions to be reduced.}

\item{eps}{is the epsilon error to be reduced according to the Johnson Loindenstrauss Lemma}
}
\value{
the dimension original information, rotation and the resulting dimension reduced matrix
}
\description{
gaussian_random_projection applies the projection onto the data.frame or matrix with the given
dimension or epsilon parameter.
}

